+++ lib/type/_message.ex
@@ -4,11 +4,11 @@ defmodule Type.Message do
   TBA
   """
 
-  @enforce_keys [:type, :target]
+  @enforce_keys [:challenge, :target]
   defstruct @enforce_keys ++ [meta: []]
 
   @type t :: %__MODULE__{
-    type:   Type.t,
+    challenge:   Type.t,
     target: Type.t,
     meta:   [
       file: Path.t,
@@ -18,7 +18,22 @@ defmodule Type.Message do
     ]
   }
 
-  def make(type, target, meta) do
-    %__MODULE__{type: type, target: target, meta: meta}
+  def make(challenge, target, meta) do
+    %__MODULE__{challenge: challenge, target: target, meta: meta}
+  end
+
+  @doc false
+  def _rebrand(:ok, _, _), do: :ok
+
+  def _rebrand({:error, message}, challenge, target) do
+    {:error, _rebrand_m(message, challenge, target)}
+  end
+
+  def _rebrand({:maybe, messages}, challenge, target) do
+    {:maybe, Enum.map(messages, &_rebrand_m(&1, challenge, target))}
+  end
+
+  def _rebrand_m(message, challenge, target) do
+    %{message | challenge: challenge, target: target}
   end
 end