+++ test/type/composite/bitstring/usable_as_test.exs
@@ -33,7 +33,7 @@ defmodule TypeTest.TypeBitString.UsableAsTest do
 
     test "is not usable as a nonzero size bitstring" do
       nonzero_bitstring = %Bitstring{size: 1, unit: 4}
-      assert {:error, %Message{type: @empty_bitstring, target: ^nonzero_bitstring}} =
+      assert {:error, %Message{challenge: @empty_bitstring, target: ^nonzero_bitstring}} =
         @empty_bitstring ~> nonzero_bitstring
     end
   end
@@ -46,22 +46,22 @@ defmodule TypeTest.TypeBitString.UsableAsTest do
     test "are not usable as other fixed-length bitstrings" do
       char2 = %Bitstring{size: 16, unit: 0}
       char1 = %Bitstring{size: 8, unit: 0}
-      assert {:error, %Message{type: ^char2, target: ^char1}} = char2 ~> char1
+      assert {:error, %Message{challenge: ^char2, target: ^char1}} = char2 ~> char1
     end
   end
 
   describe "the basic bitstring type" do
     test "is maybe usable as all other bitstrings" do
-      assert {:maybe, [%Message{type: @basic_bitstring, target: @empty_bitstring}]} =
+      assert {:maybe, [%Message{challenge: @basic_bitstring, target: @empty_bitstring}]} =
         @basic_bitstring ~> @empty_bitstring
 
-      assert {:maybe, [%Message{type: @basic_bitstring, target: @basic_binary}]} =
+      assert {:maybe, [%Message{challenge: @basic_bitstring, target: @basic_binary}]} =
         @basic_bitstring ~> @basic_binary
 
-      assert {:maybe, [%Message{type: @basic_bitstring, target: %Bitstring{size: 7, unit: 2}}]} =
+      assert {:maybe, [%Message{challenge: @basic_bitstring, target: %Bitstring{size: 7, unit: 2}}]} =
         @basic_bitstring ~> %Bitstring{size: 7, unit: 2}
 
-      assert {:maybe, [%Message{type: @basic_bitstring, target: %Bitstring{size: 3, unit: 1}}]} =
+      assert {:maybe, [%Message{challenge: @basic_bitstring, target: %Bitstring{size: 3, unit: 1}}]} =
         @basic_bitstring ~> %Bitstring{size: 3, unit: 1}
     end
   end
@@ -78,15 +78,15 @@ defmodule TypeTest.TypeBitString.UsableAsTest do
         @basic_binary ~> %Bitstring{size: 4, unit: 4}
 
       # not known if there are enough multiples to get there.
-      assert {:maybe, [%Message{type: @basic_binary, target: %Bitstring{size: 8, unit: 0}}]} =
+      assert {:maybe, [%Message{challenge: @basic_binary, target: %Bitstring{size: 8, unit: 0}}]} =
         @basic_binary ~> %Bitstring{size: 8, unit: 0}
       # competely strange factors that sometimes multiple to 8
-      assert {:maybe, [%Message{type: @basic_binary, target: %Bitstring{size: 3, unit: 5}}]} =
+      assert {:maybe, [%Message{challenge: @basic_binary, target: %Bitstring{size: 3, unit: 5}}]} =
         @basic_binary ~> %Bitstring{size: 3, unit: 5}
     end
 
     test "is never usable as something which can never be a non-multiple-of-8" do
-      assert {:error, %Message{type: @basic_binary, target: %Bitstring{size: 3, unit: 4}}} =
+      assert {:error, %Message{challenge: @basic_binary, target: %Bitstring{size: 3, unit: 4}}} =
         @basic_binary ~> %Bitstring{size: 3, unit: 4}
     end
   end
@@ -96,7 +96,7 @@ defmodule TypeTest.TypeBitString.UsableAsTest do
       bs_1 = %Bitstring{size: 13, unit: 3}
       bs_2 = %Bitstring{size: 7, unit: 9}
 
-      assert {:maybe, [%Message{type: ^bs_1, target: ^bs_2}]} = bs_1 ~> bs_2
+      assert {:maybe, [%Message{challenge: ^bs_1, target: ^bs_2}]} = bs_1 ~> bs_2
     end
 
     test "same unit, different sizes." do
@@ -110,7 +110,7 @@ defmodule TypeTest.TypeBitString.UsableAsTest do
       bs_1 = %Bitstring{size: 5, unit: 7}
       bs_2 = %Bitstring{size: 9, unit: 2}
 
-      assert {:maybe, [%Message{type: ^bs_1, target: ^bs_2}]} = bs_1 ~> bs_2
+      assert {:maybe, [%Message{challenge: ^bs_1, target: ^bs_2}]} = bs_1 ~> bs_2
     end
   end
 
@@ -125,17 +125,17 @@ defmodule TypeTest.TypeBitString.UsableAsTest do
   test "bitstrings generally are not usable as other types" do
     targets = TypeTest.Targets.except([@empty_bitstring, type(<<>>), "foo", <<0::7>>])
     Enum.each(targets, fn target ->
-      assert {:error, %Message{type: @empty_bitstring, target: ^target}} =
+      assert {:error, %Message{challenge: @empty_bitstring, target: ^target}} =
         (@empty_bitstring ~> target)
     end)
 
     Enum.each(targets, fn target ->
-      assert {:error, %Message{type: @basic_binary, target: ^target}} =
+      assert {:error, %Message{challenge: @basic_binary, target: ^target}} =
         (@basic_binary ~> target)
     end)
 
     Enum.each(targets, fn target ->
-      assert {:error, %Message{type: @basic_bitstring, target: ^target}} =
+      assert {:error, %Message{challenge: @basic_bitstring, target: ^target}} =
         (@basic_bitstring ~> target)
     end)
   end